+++ src/app/(sidebar)/my-recruit/[id]/components/DetailHeader.tsx
@@ -3,29 +3,22 @@
 import { useOutsideClick } from '@/hooks/useOutsideClick';
 import { Button, Icon } from '@/system/components';
 import { color } from '@/system/token/color';
-import { cn } from '@/utils';
-import { ChangeEvent, useRef, useState } from 'react';
+import { useRef, useState } from 'react';
 import { useGetRecruitById } from '../api/useGetRecruitById';
-import { usePatchTitle } from '../api/usePatchTitle';
 import { ApplicationStatus } from './ApplicationStatus';
 import SemesterSelector from './SemesterSelector';
 import TextBubble from './TextBubble';
+import TitleInput from './TitleInput';
 
 export default function DetailHeader({ recruitId }: { recruitId: string }) {
-  const { mutate: patchTitle } = usePatchTitle();
   const { data: recruitInfoById } = useGetRecruitById(recruitId);
 
   const tooltipRef = useRef<HTMLDivElement>(null);
-  const [title, setTitle] = useState(recruitInfoById?.title || '');
   const [isFocused, setIsFocused] = useState(false);
   const [isHovered, setIsHovered] = useState(false);
   const [isLinked, setIsLinked] = useState(false);
 
   useOutsideClick(tooltipRef, () => setIsLinked(false));
-
-  const handleTitleChange = (e: ChangeEvent<HTMLInputElement>) => {
-    setTitle(e.target.value);
-  };
 
   const handleSetLink = () => {
     setIsLinked(true);
@@ -47,24 +40,12 @@ export default function DetailHeader({ recruitId }: { recruitId: string }) {
     }, 2000);
   };
 
-  const handleUpdateTitle = () => {
-    setIsFocused(false);
-    patchTitle({ newTitle: title, id: recruitId });
-  };
-
   return (
     <div className="flex gap-[0.75rem] items-center">
       <SemesterSelector recruitId={recruitId} season={recruitInfoById?.season || ''} />
       <ApplicationStatus recruitId={recruitId} status={recruitInfoById?.recruitStatus || ''} />
-      <input
-        className={cn(
-          'pl-[6px] mr-[6px] rounded-[6px] text-neutral-95 text-heading1 font-bold border-none hover:bg-neutral-3 focus:outline-none focus:ring-2 focus:ring-mint-20 focus:ring-offset-2  focus:hover:bg-white',
-        )}
-        onChange={handleTitleChange}
-        value={title}
-        onFocus={() => setIsFocused(true)}
-        onBlur={handleUpdateTitle}
-      />
+
+      <TitleInput recruitInfoById={recruitInfoById} recruitId={recruitId} setIsFocused={setIsFocused} />
 
       <div ref={tooltipRef} className="relative">
         {!isFocused && (
